FareShare/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
pdfplumber==0.11.4
pypdf==4.3.1
sqlalchemy==2.0.35
alembic==1.13.3
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
reportlab==4.2.2
pillow==10.4.0
aiofiles==24.1.0
pydantic-settings==2.4.0
